LayoutTests/accessibility/mac/ancestor-attributes-expected.txt


This tests attributes for looking up editable and focusable ancestors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.isEqual(focusableAncestor) is true
PASS input.isEqual(editableAncestor) is true
PASS editable.isEqual(highestAncestor) is true
PASS successfullyParsed is true

TEST COMPLETE
